[#2]rename: Product -> Post 이름 변경 및 파일위치 이동
