+++ src/main/java/com/toyproject/usedauction/domain/post/OrderPriceType.java
@@ -1,4 +1,4 @@
-package com.toyproject.usedauction.domain.product;
+package com.toyproject.usedauction.domain.post;
 
 public enum OrderPriceType {
